Refactored Test Code #CHANGE: Refactored component execution order test code to be more modular and readable.
