+++ Test/Core/Components/ComponentExecutionOrderTest.cs
@@ -155,74 +155,6 @@ namespace Duality.Tests.Components
 
 	namespace ExecutionOrderTest
 	{
-		[Flags]
-		public enum EventType
-		{
-			None = 0x0,
-
-			Update = 0x1,
-			Activate = 0x2,
-			Deactivate = 0x4,
-
-			All = Update | Activate | Deactivate
-		}
-		public class EventOrderLog
-		{
-			private EventType eventFilter = EventType.All;
-			private List<TestComponent> eventOrder = new List<TestComponent>();
-
-			public EventType EventFilter
-			{
-				get { return this.eventFilter; }
-				set { this.eventFilter = value; }
-			}
-			public IEnumerable<TestComponent> EventOrder
-			{
-				get { return this.eventOrder; }
-			}
-
-			public void Clear()
-			{
-				this.eventOrder.Clear();
-			}
-			public void Notify(EventType eventType, TestComponent receiver)
-			{
-				if ((this.eventFilter & eventFilter) == EventType.None) return;
-				this.eventOrder.Add(receiver);
-			}
-		}
-		public class TestComponent : Component, ICmpUpdatable, ICmpInitializable
-		{
-			private EventOrderLog eventLog;
-
-			public EventOrderLog EventLog
-			{
-				get { return this.eventLog; }
-				set { this.eventLog = value; }
-			}
-
-			private void NotifyEvent(EventType type)
-			{
-				if (this.eventLog == null) return;
-				this.eventLog.Notify(type, this);
-			}
-
-			void ICmpUpdatable.OnUpdate()
-			{
-				this.NotifyEvent(EventType.Update);
-			}
-			void ICmpInitializable.OnInit(Component.InitContext context)
-			{
-				if (context == InitContext.Activate)
-					this.NotifyEvent(EventType.Activate);
-			}
-			void ICmpInitializable.OnShutdown(Component.ShutdownContext context)
-			{
-				if (context == ShutdownContext.Deactivate)
-					this.NotifyEvent(EventType.Deactivate);
-			}
-		}
-
 		public class TestComponentA1 : TestComponent { }
 		public class TestComponentA2 : TestComponent { }
 		public class TestComponentA3 : TestComponent { }
