+++ Test/Core/DualityTests.csproj
@@ -76,6 +76,9 @@
     <Compile Include="Cloning\SpecificCloningTest.cs" />
     <Compile Include="Components\ComponentExecutionOrderTest.cs" />
     <Compile Include="Components\ComponentRequirementMapTest.cs" />
+    <Compile Include="Components\ExecutionOrderTest\EventOrderLog.cs" />
+    <Compile Include="Components\ExecutionOrderTest\EventType.cs" />
+    <Compile Include="Components\ExecutionOrderTest\TestComponent.cs" />
     <Compile Include="Components\RigidBodyTest.cs" />
     <Compile Include="ExtMethodsPixelData.cs" />
     <Compile Include="Math\Vector4Test.cs" />
